Fix generation of write off for multiple moves with use of tolerance
Fix application of discount with cash rounding

If a cash rounding is applied, there can be two move lines for the
tax as a rounding one is added automatically.
The condition to store the financial discount must therefore check
if all the tax lines share the same tax. The financial discount
will then be linked with the bigger amount tax line in order to
generate the proper write off in the reconciliation.

Define journal_id key in write_off to avoid error in _prepare_widget_writeoff_vals

This is necessary due to the new call to _prepare_widget_writeoff_vals with the
enterprise reconciliation widget added in following commit:
https://github.com/odoo/odoo/pull/85475/commits/1c43c5315ba6e78e3cb818ff5131ec441ab06dfe

Fix multicurrency in account payment register

Add lots of tests to cover most of the use cases

Fix floats comparisons

Fix error on Monetary fields without value accessed by SQL

Without the default value to 0.0, Monetary fields will have a NULL
value in the database if not set explicitely.
When reading such values through the ORM, it will be defaulted to
0.0, but if we read it through a direct SQL query, the NULL values
will be set a None in the results, what could raise errors when
doing calculations on it.

Add default value on the fields and use COALESCE in SQL query should
ensure we never get such errors again.

Refactor application of taxes in reconciliation write off

account_financial_discount: Improve manifest